Adding last of the separated resolvers.
